refactor: 공통 함수 dateUtils로 분리
